+++ packages/toolkit/create/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @modern-js/create
 
+## 3.7.0
+
+## 3.6.0
+
 ## 3.5.0
 
 ### BleedingDev Strict Effect API Migration Notes